requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
tensorflow>=2.13.0
pandas>=2.0.0
numpy>=1.24.0
boto3>=1.34.0
requests>=2.31.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
scipy>=1.11.0
pyportfolioopt>=1.5.0
cachetools>=5.3.0
